Feature/handle edge case request artwork service

+++ src/main/java/com/github/khangzxrr/domain/RequestBid.java
@@ -24,7 +24,7 @@ public class RequestBid implements Serializable {
     private String description;
 
     @Column(name = "price")
-    private Long price;
+    private Double price;
 
     @Column(name = "duration")
     private Long duration;
@@ -68,16 +68,16 @@ public class RequestBid implements Serializable {
         this.description = description;
     }
 
-    public Long getPrice() {
+    public Double getPrice() {
         return this.price;
     }
 
-    public RequestBid price(Long price) {
+    public RequestBid price(Double price) {
         this.setPrice(price);
         return this;
     }
 
-    public void setPrice(Long price) {
+    public void setPrice(Double price) {
         this.price = price;
     }
